requirements.txt
asyncio
websockets
python-dotenv
python-binance